+++ src/identifiable.rs
@@ -19,8 +19,37 @@ pub const STRING_ID: u32 = 0xb5286e24;
 pub const VECTOR_ID: u32 = 0x1cb5c415;
 
 
+const BOOL_IDS: &'static [u32] = &[BOOL_TRUE_ID, BOOL_FALSE_ID];
+const INT_IDS: &'static [u32] = &[INT_ID];
+const LONG_IDS: &'static [u32] = &[LONG_ID];
+const DOUBLE_IDS: &'static [u32] = &[DOUBLE_ID];
+const STRING_IDS: &'static [u32] = &[STRING_ID];
+const VECTOR_IDS: &'static [u32] = &[VECTOR_ID];
+
+
 /// A trait for a Rust data structure that can have an id.
 pub trait Identifiable {
+    /// Get all possible ids (known at compile time) of an identifiable type.
+    ///
+    /// This is most useful for enums where each variant has its own id.
+    ///
+    /// __Implementation note__: this method **should** return a slice of
+    /// **non-duplicate** values, i.e. the slice must effectively be a set with
+    /// contents known at compile-time.
+    /// Currently, there is no way to enforce this restriction using the
+    /// language itself.
+    /// This restriction is marked as **should** because it potentially only
+    /// alters the behavior of counting all ids a type is question has, but not
+    /// checking if an arbitrary id is in this set - in this case, a reliable
+    /// counting routine must traverse the whole slice to eliminate duplicates.
+    /// Unfortunately, this worsens the time complexity from O(*1*) to O(*n*),
+    /// but for everyday use-case this is fine since product types we usually
+    /// use are relatively small to make this a big concern.
+    ///
+    /// Compatibility note: will probably be replaced by an associated constant
+    /// after bumping minimum supported Rust version to 1.20.
+    fn all_type_ids() -> &'static [u32];
+
     /// Get id of a value of an identifiable type.
     ///
     /// Its signature is made `(&self) -> i32`, not `() -> i32` because of enum
@@ -38,6 +67,10 @@ pub trait Identifiable {
 
 
 impl<'a, T: Identifiable> Identifiable for &'a T {
+    fn all_type_ids() -> &'static [u32] {
+        T::all_type_ids()
+    }
+
     fn type_id(&self) -> u32 {
         (*self).type_id()
     }
@@ -48,6 +81,10 @@ impl<'a, T: Identifiable> Identifiable for &'a T {
 }
 
 impl<T: Identifiable> Identifiable for Box<T> {
+    fn all_type_ids() -> &'static [u32] {
+        T::all_type_ids()
+    }
+
     fn type_id(&self) -> u32 {
         (**self).type_id()
     }
@@ -59,6 +96,10 @@ impl<T: Identifiable> Identifiable for Box<T> {
 
 #[cfg_attr(feature = "cargo-clippy", allow(match_bool))]  // match looks better here
 impl Identifiable for bool {
+    fn all_type_ids() -> &'static [u32] {
+        BOOL_IDS
+    }
+
     fn type_id(&self) -> u32 {
         match *self {
             false => BOOL_FALSE_ID,
@@ -77,11 +118,15 @@ impl Identifiable for bool {
 
 
 macro_rules! impl_identifiable_for_primitives {
-    ($($type:ty => $id_value:expr,)*) => {
+    ($($type:ty => ($all_ids:expr, $id_of_value:expr),)*) => {
         $(
             impl Identifiable for $type {
+                fn all_type_ids() -> &'static [u32] {
+                    $all_ids
+                }
+
                 fn type_id(&self) -> u32 {
-                    $id_value
+                    $id_of_value
                 }
 
                 fn enum_variant_id(&self) -> Option<&'static str> {
@@ -93,23 +138,27 @@ macro_rules! impl_identifiable_for_primitives {
 }
 
 impl_identifiable_for_primitives! {
-    i8  => INT_ID,
-    i16 => INT_ID,
-    i32 => INT_ID,
-    i64 => LONG_ID,
+    i8  => (INT_IDS,  INT_ID),
+    i16 => (INT_IDS,  INT_ID),
+    i32 => (INT_IDS,  INT_ID),
+    i64 => (LONG_IDS, LONG_ID),
 
-    u8  => INT_ID,
-    u16 => INT_ID,
-    u32 => INT_ID,
-    u64 => LONG_ID,
+    u8  => (INT_IDS,  INT_ID),
+    u16 => (INT_IDS,  INT_ID),
+    u32 => (INT_IDS,  INT_ID),
+    u64 => (LONG_IDS, LONG_ID),
 
-    f32 => DOUBLE_ID,
-    f64 => DOUBLE_ID,
+    f32 => (DOUBLE_IDS, DOUBLE_ID),
+    f64 => (DOUBLE_IDS, DOUBLE_ID),
 
-    String => STRING_ID,
+    String => (STRING_IDS, STRING_ID),
 }
 
 impl<'a> Identifiable for &'a str {
+    fn all_type_ids() -> &'static [u32] {
+        STRING_IDS
+    }
+
     fn type_id(&self) -> u32 {
         STRING_ID
     }
@@ -120,6 +169,10 @@ impl<'a> Identifiable for &'a str {
 }
 
 impl<T> Identifiable for Vec<T> {
+    fn all_type_ids() -> &'static [u32] {
+        VECTOR_IDS
+    }
+
     fn type_id(&self) -> u32 {
         VECTOR_ID
     }